Class/Test/CustomerSupportTest.java
package Test;

import org.junit.Before;
import org.junit.Test;
import static org.junit.Assert.*;
import Class.CustomerSupport;

public class CustomerSupportTest {
    private CustomerSupport customerSupport;

    @Before
    public void setUp() {
        // Initialize CustomerSupport with test data
        customerSupport = new CustomerSupport(1, "John Doe");
    }

    @Test
    public void testPrioritizeIssues() {
        // This test assumes the prioritizeIssues method affects some internal state or logs action
        customerSupport.prioritizeIssues();
        // Here, you would verify that the prioritization action was taken
        // This step is conceptual since the actual method prints to the console
        assertTrue("Issues should be prioritized", true);
    }

    @Test
    public void testProvideAssistance() {
        // This test assumes provideAssistance logs the assistance details or changes state
        customerSupport.provideAssistance("Assistance for issue XYZ");
        // Verify the assistance was provided
        // This assertion is conceptual. You might check a log or a state change in a real test
        assertTrue("Assistance should be provided", true);
    }

    @Test
    public void testVerifyInformation() {
        // Assuming verifyInformation method logs information or updates state
        customerSupport.verifyInformation("Customer Info ABC");
        // Verify the information was verified
        // This is a conceptual placeholder assertion
        assertTrue("Information should be verified", true);
    }
}
